requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
pydantic
pydantic-settings
jinja2
httpx
python-slugify
pytest
pytest-asyncio
